fix android compiling tools
- Added -D_STRINGS_H_ to `$CFLAGS` because `stdlib.h` includes
`strings.h` which has duplicates symbols when used with `strings.h`
- Removed soft dependency of tools/Makefile:makescripts on
common_make_rules creating `$TOP/bin` for it when not cross-compiling.
Now it just simply uses `$BINDIR` which is set to `$TOP/bin` when not
cross-compiling. This is parity and makes a lot more sense if you're
gonna package this.